Move client trpc config to `shared`
