+++ apps/dashboard/src/pages/index.tsx
@@ -1,4 +1,4 @@
-import { trpc } from '../app/trpc'
+import { trpc } from '../shared/api/trpc'
 export default function IndexPage() {
   const hello = trpc.hello.useQuery({ text: 'client' })
   if (!hello.data) {
