+++ apps/dashboard/src/shared/api/trpc.ts
@@ -1,6 +1,6 @@
 import { httpBatchLink } from '@trpc/client'
 import { createTRPCNext } from '@trpc/next'
-import type { AppRouter } from '../server/routers/_app'
+import type { AppRouter } from '../../server/routers/_app'
 function getBaseUrl() {
   if (typeof window !== 'undefined')
     // browser should use relative path
